+++ app/schemas/product.py
@@ -1,12 +1,11 @@
 from pydantic import BaseModel, ConfigDict, Field
-from decimal import Decimal
 from datetime import datetime
 
 
 class ProductCreate(BaseModel):
     name: str = Field(min_length=3, max_length=125)
     description: str | None = Field(default=None, max_length=1500)
-    price: Decimal = Field(gt=0)
+    price: int = Field(gt=0)
     quantity: int = Field(ge=0)
     category_id: int
 
@@ -16,7 +15,7 @@ class ProductCreate(BaseModel):
 class ProductUpdate(BaseModel):
     name: str | None = Field(default=None, min_length=3, max_length=125)
     description: str | None = Field(default=None, max_length=1500)
-    price: Decimal | None = Field(default=None, gt=0)
+    price: int | None = Field(default=None, gt=0)
     quantity: int | None = Field(default=None, ge=0)
     category_id: int | None = Field(default=None)
 
@@ -27,7 +26,7 @@ class ProductResponse(BaseModel):
     id: int
     name: str
     description: str | None
-    price: Decimal
+    price: int
     quantity: int
     category_id: int
     created_at: datetime
